services/grades_service/requirements.txt
Django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-filter==23.5
whitenoise==6.4.0
gunicorn==20.1.0 